Bugfix for selected class with padTime enabled
Using the timepicker with the prop "padTime" enabled I figured out a bug that prevent the selected picker item to have the right "selected" class because the condition results false because of the triple equal (===) comparison between string and integer. Using double equal comparison for hours, minutes and seconds it results now fixed.